% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asrproject.R
\name{get_tree}
\alias{get_tree}
\title{Test package.}
\usage{
get_tree(phy, params_file, q_matrix_file)
}
\arguments{
\item{phy}{TODO has tip.state}

\item{params_file}{TODO}

\item{q_matrix_file}{TODO}
}
\value{
TODO phylo
}
\description{
Test package.
}
